refactor: centralize User module states and hooks
Restructure the User module folder by centralizing all state management
and hooks related to the User module. This change improves maintainability
and provides a clearer organization for User-related states and hooks.
