+++ components/user/userSessionGroupEffect/userSessionEffect/index.tsx
@@ -1,5 +1,5 @@
+import { useUserSession } from '@user/user.hooks';
 import { useEffect } from 'react';
-import { useUserSession } from './userSessionEffect.hooks';
 
 export const UserSessionEffect = () => {
   const userSession = useUserSession();
