+++ jest.config.js
@@ -29,7 +29,7 @@ const customJestConfig = {
     '^@stateLogics/(.*)$': ['<rootDir>/lib/stateLogics/$1'],
     '^@data/(.*)$': ['<rootDir>/lib/data/$1'],
     '^@lib/(.*)$': ['<rootDir>/lib/$1'],
-    '^@users/(.*)$': ['<rootDir>/components/users/$1'],
+    '^@user/(.*)$': ['<rootDir>/components/user/$1'],
     '^@layouts/(.*)$': ['<rootDir>/components/layouts/$1'],
     '^@buttons/(.*)$': ['<rootDir>/components/ui/buttons/$1'],
     '^@dropdowns/(.*)$': ['<rootDir>/components/ui/dropdowns/$1'],
